Hero buy and get logic moved to hero controller Created rest methods for hero information Some DTOs associated with ws moved to separate package
