--- src/main/kotlin/com/vakzu/musicwars/dto/OnlineMessage.kt
@@ -1,5 +1,0 @@
-package com.vakzu.musicwars.dto
-
-import com.vakzu.musicwars.dto.websocket.CommandType
-
-class OnlineMessage(val type: CommandType, val userId: Int, val username: String)
