--- src/main/kotlin/com/vakzu/musicwars/dto/ReadyResponse.kt
@@ -1,5 +1,0 @@
-package com.vakzu.musicwars.dto
-
-import com.vakzu.musicwars.dto.websocket.CommandType
-
-class ReadyResponse(val type: CommandType, val userId: Int)
