+++ src/main/kotlin/com/vakzu/musicwars/dto/websocket/OnlineMessage.kt
@@ -1,0 +1,3 @@
+package com.vakzu.musicwars.dto.websocket
+
+class OnlineMessage(val type: CommandType, val userId: Int, val username: String)
