+++ src/main/kotlin/com/vakzu/musicwars/dto/websocket/ReadyResponse.kt
@@ -1,0 +1,3 @@
+package com.vakzu.musicwars.dto.websocket
+
+class ReadyResponse(val type: CommandType, val userId: Int)
