+++ src/main/kotlin/com/vakzu/musicwars/dto/websocket/SetReadyRequest.kt
@@ -1,5 +1,3 @@
-package com.vakzu.musicwars.dto
-
-import com.vakzu.musicwars.dto.websocket.CommandType
+package com.vakzu.musicwars.dto.websocket
 
 class SetReadyRequest(val commandType: CommandType, val characterId: Int, val songId: Int, val effectId: Int)
